Add files via upload
This Jupyter notebook provides an introduction to web scraping and a step-by-step guide on how to use the requests and BeautifulSoup libraries in Python to extract customer reviews for the iPhone 11 from Flipkart.com. It also includes an overview of the code used in the scraping process and explanations of each step. This notebook is a great starting point for anyone interested in learning about web scraping and data extraction using Python.